requirements.txt
#LangChain Core
langchain
langchain-core
langchain-community

#OpenAI Integration
langchain-openai
openai

#Anthropic Integration
langchain-anthropic

#Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Ollama Integration
langchain-ollama

#Environment Variable Management
python-dotenv

#Machine Learning Utilities
numpy
scikit-learn

#Streamlit for UI
streamlit

#MySQL Connector
mysql-connector-python

#Vector Store
faiss-cpu

#PyTorch for Transformers
torch